SensorTag connection Fix

+++ MSP430_HCI/hal_types.h
@@ -56,8 +56,8 @@
 typedef signed   char   int8;     //!< Signed 8 bit integer
 typedef unsigned char   uint8;    //!< Unsigned 8 bit integer
 
-typedef signed   short  int16;    //!< Signed 16 bit integer
-typedef unsigned short  uint16;   //!< Unsigned 16 bit integer
+typedef signed   int  int16;    //!< Signed 16 bit integer
+typedef unsigned int  uint16;   //!< Unsigned 16 bit integer
 
 
 typedef  char  bool;     //!< Boolean data type
